The great "(self)" purge of 2023
An explanation: Qt has this concept of parent-child relationships, where object A can be the parent of object B. If A is the parent of B, then B won't get garbage-collected so long as A exists. But that's usually unnecessary because Python already does reference counting. So if A has a reference to B, then B will stick around as long as A does regardless of parent-child stuff. I'm removing the few parent-child relationships that are left in the code to tidy things up.